[WEB-3395] - Moved CrmUsersService to UsersModule
