+++ src/admin/campaigns/adminCampaigns.service.ts
@@ -5,7 +5,7 @@ import { Campaign, Prisma } from '@prisma/client'
 import { EmailService } from 'src/email/email.service'
 import { getUserFullName } from 'src/users/util/users.util'
 import { EmailTemplateNames } from 'src/email/email.types'
-import { UsersService } from 'src/users/users.service'
+import { UsersService } from 'src/users/services/users.service'
 import { CampaignsService } from 'src/campaigns/services/campaigns.service'
 import { AdminP2VService } from '../services/adminP2V.service'
 import { CampaignWith, OnboardingStep } from 'src/campaigns/campaigns.types'
