+++ src/authentication/auth-strategies/JwtAuth.strategy.ts
@@ -2,7 +2,7 @@ import { ExtractJwt, Strategy } from 'passport-jwt'
 import { PassportStrategy } from '@nestjs/passport'
 import { Injectable, UnauthorizedException } from '@nestjs/common'
 import { JwtPayload } from 'jsonwebtoken'
-import { UsersService } from '../../users/users.service'
+import { UsersService } from '../../users/services/users.service'
 
 @Injectable()
 export class JwtAuthStrategy extends PassportStrategy(Strategy) {
