+++ src/authentication/authentication.controller.ts
@@ -15,7 +15,7 @@ import { ZodValidationPipe } from 'nestjs-zod'
 import { ReadUserOutputSchema } from '../users/schemas/ReadUserOutput.schema'
 import { RecoverPasswordSchema } from './schemas/RecoverPasswordEmail.schema'
 import { SetPasswordEmailSchema } from './schemas/SetPasswordEmail.schema'
-import { UsersService } from 'src/users/users.service'
+import { UsersService } from 'src/users/services/users.service'
 import { EmailService } from 'src/email/email.service'
 import { ResetPasswordSchema } from './schemas/ResetPassword.schema'
 import { CampaignsService } from 'src/campaigns/services/campaigns.service'
