+++ src/crm/crmCampaigns.service.ts
@@ -15,8 +15,8 @@ import { SlackService } from '../shared/services/slack.service'
 import { Campaign, User } from '@prisma/client'
 import { getUserFullName } from '../users/util/users.util'
 import { formatDateForCRM, getCrmP2VValues } from './util/cms.util'
-import { CrmUsersService } from './crmUsers.service'
-import { UsersService } from '../users/users.service'
+import { CrmUsersService } from '../users/services/crmUsers.service'
+import { UsersService } from '../users/services/users.service'
 import { AssociationSpecAssociationCategoryEnum } from '@hubspot/api-client/lib/codegen/crm/associations/v4/models/AssociationSpec'
 import { AssociationTypes } from '@hubspot/api-client'
 import { AiChatService } from '../campaigns/ai/chat/aiChat.service'
