+++ src/payments/stripe/stripeEvents.service.ts
@@ -10,7 +10,7 @@ import { StripeSingleton } from './stripe.service'
 import { WebhookEventType } from '../payments.types'
 import Stripe from 'stripe'
 import { CampaignsService } from '../../campaigns/services/campaigns.service'
-import { UsersService } from '../../users/users.service'
+import { UsersService } from '../../users/services/users.service'
 import { SlackService } from '../../shared/services/slack.service'
 import { Campaign, User } from '@prisma/client'
 import { DateFormats, formatDate } from '../../shared/util/date.util'
