+++ src/users/services/crmUsers.service.ts
@@ -6,14 +6,14 @@ import {
   Logger,
 } from '@nestjs/common'
 import { Campaign, User } from '@prisma/client'
-import { UsersService } from '../users/users.service'
-import { CampaignsService } from '../campaigns/services/campaigns.service'
-import { getMidnightForDate } from '../shared/util/date.util'
-import { HubspotService } from './hubspot.service'
-import { CRMContactProperties } from './crm.types'
+import { UsersService } from './users.service'
+import { CampaignsService } from '../../campaigns/services/campaigns.service'
+import { getMidnightForDate } from '../../shared/util/date.util'
+import { HubspotService } from '../../crm/hubspot.service'
+import { CRMContactProperties } from '../../crm/crm.types'
 import { HttpService } from '@nestjs/axios'
 import { lastValueFrom } from 'rxjs'
-import { SlackService } from '../shared/services/slack.service'
+import { SlackService } from '../../shared/services/slack.service'
 import { Headers, MimeTypes } from 'http-constants-ts'
 import { AxiosError, isAxiosError } from 'axios'
